Extract common custom-node-import checker
