+++ src/extractors/customTransComponent.ts
@@ -1,51 +1,11 @@
-import { dirname, isAbsolute, relative, sep } from 'path';
 import * as BabelTypes from '@babel/types';
 import * as BabelCore from '@babel/core';
 import { getCommentHintForPath, CommentHint } from '../comments';
-import { referencesImport } from './commons';
 import extractTransComponent from './transComponent';
 import { ExtractedKey } from '../keys';
 import { Config } from '../config';
 
-/**
- * Check whether a given JSXElement is a custom Trans component.
- *
- * @param absoluteCustomTransComponents: list of possible Trans components
- *   with their source modules and import names.
- * @param path: node path to check
- * @returns true if the given element is a custom `Trans` component.
- */
-function isCustomTransComponent(
-  absoluteCustomTransComponents: Config['cache']['absoluteCustomTransComponents'],
-  path: BabelCore.NodePath<BabelTypes.JSXElement>,
-): boolean {
-  const openingElementName = path.get('openingElement').get('name');
-  return absoluteCustomTransComponents.some(
-    ([sourceModule, importName]): boolean => {
-      if (isAbsolute(sourceModule)) {
-        let relativeSourceModulePath = relative(
-          dirname(path.state.filename),
-          sourceModule,
-        );
-        if (!relativeSourceModulePath.startsWith('.')) {
-          relativeSourceModulePath = '.' + sep + relativeSourceModulePath;
-        }
-
-        // Absolute path to the source module, let's try a relative path first.
-        if (
-          referencesImport(
-            openingElementName,
-            relativeSourceModulePath,
-            importName,
-          )
-        ) {
-          return true;
-        }
-      }
-      return referencesImport(openingElementName, sourceModule, importName);
-    },
-  );
-}
+import { isCustomImportedNode } from './commons';
 
 /**
  * Extract custom Trans components.
@@ -61,7 +21,11 @@ export default function extractCustomTransComponent(
 ): ExtractedKey[] {
   if (getCommentHintForPath(path, 'DISABLE', commentHints)) return [];
   if (
-    !isCustomTransComponent(config.cache.absoluteCustomTransComponents, path)
+    !isCustomImportedNode(
+      config.cache.absoluteCustomTransComponents,
+      path,
+      path.get('openingElement').get('name'),
+    )
   )
     return [];
   return extractTransComponent(path, config, commentHints, true);
